Refactor pkg/util/server to avoid cortexpb import in downstream projects vendoring logql
