+++ pkg/util/server/middleware.go
@@ -1,22 +1,10 @@
 package server
 
 import (
-	"context"
 	"net/http"
-	"regexp"
 
 	"github.com/weaveworks/common/httpgrpc"
 	"github.com/weaveworks/common/middleware"
-)
-
-// NOTE(kavi): Why new type?
-// Our linter won't allow to use basic types like string to be used as key in context.
-type ctxKey string
-
-var (
-	QueryTagsHTTPHeader ctxKey = "X-Query-Tags"
-	safeQueryTags              = regexp.MustCompile("[^a-zA-Z0-9-=, ]+") // only alpha-numeric, ' ', ',', '=' and `-`
-
 )
 
 // NewPrepopulateMiddleware creates a middleware which will parse incoming http forms.
@@ -39,22 +27,6 @@ func ResponseJSONMiddleware() middleware.Interface {
 	return middleware.Func(func(next http.Handler) http.Handler {
 		return http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
 			w.Header().Set("Content-Type", "application/json; charset=UTF-8")
-			next.ServeHTTP(w, req)
-		})
-	})
-}
-
-func ExtractQueryTagsMiddleware() middleware.Interface {
-	return middleware.Func(func(next http.Handler) http.Handler {
-		return http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
-			ctx := req.Context()
-			tags := req.Header.Get(string(QueryTagsHTTPHeader))
-			tags = safeQueryTags.ReplaceAllString(tags, "")
-
-			if tags != "" {
-				ctx = context.WithValue(ctx, QueryTagsHTTPHeader, tags)
-				req = req.WithContext(ctx)
-			}
 			next.ServeHTTP(w, req)
 		})
 	})
